chore: reorganize files in a directory
